+++ src/component/Theme/Paper/LightPaper.tsx
@@ -1,11 +1,11 @@
 import React from "react";
-import { PaperProps } from "@mantine/core";
 import { MergedPaper } from "./MergedPaper";
+import { ThemePaperProps } from "../../../util";
 
 /**
  * A light-themed [Paper](https://mantine.dev/core/paper) component.
  */
-export function LightPaper({ withBorder = true, ...other }: PaperProps) {
+export function LightPaper({ withBorder = true, ...other }: ThemePaperProps) {
     return (
         <MergedPaper withBorder={withBorder} {...other} themeSx={theme => ({
             "background": theme.colors.gray[1]