+++ src/component/Theme/Paper/MergedPaper.tsx
@@ -1,21 +1,21 @@
 import React from "react";
-import { Paper, PaperProps } from "@mantine/core";
+import { Paper } from "@mantine/core";
 import { ThemeSx } from "../interface/ThemeSx";
-import { mergeSx } from "../../../util";
+import { mergeSx, ThemePaperProps } from "../../../util";
 
 /**
  * Options for the {@link MergedPaper} component.
  */
-interface MergedPaperProps extends PaperProps, ThemeSx {}
+interface MergedPaperProps extends ThemePaperProps, ThemeSx {}
 
 /**
  * A standardized [Paper](https://mantine.dev/core/paper) component.
  *
  * Intended for use within {@link DarkPaper} or {@link LightPaper}.
  */
-export function MergedPaper({ children, themeSx, shadow = "xs", p = "md", sx, ...other }: MergedPaperProps) {
+export function MergedPaper({ children, themeSx, id, shadow = "xs", p = "md", sx, ...other }: MergedPaperProps) {
     return (
-        <Paper shadow={shadow} p={p} sx={mergeSx(themeSx, sx)} {...other}>
+        <Paper id={id} shadow={shadow} p={p} sx={mergeSx(themeSx, sx)} {...other}>
             {children}
         </Paper>
     );