+++ src/component/Theme/Paper/ThemePaper.tsx
@@ -1,12 +1,13 @@
 import React from "react";
-import { useMantineColorScheme, PaperProps } from "@mantine/core";
+import { useMantineColorScheme } from "@mantine/core";
 import { DarkPaper } from "./DarkPaper";
 import { LightPaper } from "./LightPaper";
+import { ThemePaperProps } from "../../../util";
 
 /**
  * A [Paper](https://mantine.dev/core/paper) component that automatically changes with the theme.
  */
-export function ThemePaper({ ...other }: PaperProps) {
+export function ThemePaper({ ...other }: ThemePaperProps) {
     const { colorScheme } = useMantineColorScheme();
 
     return (