+++ src/util/index.ts
@@ -1,5 +1,6 @@
 export * from "./interface/SxOperations";
 export * from "./interface/caption/SharedCaptionProps";
+export * from "./interface/themePaper/ThemePaperProps";
 export * from "./interface/video/SharedVideoProps";
 export * from "./notification/error";
 export * from "./style/merge";